adding check before deleting

+++ src/modules/exams/services/DeleteExam.service.ts
@@ -1,3 +1,4 @@
+import { ObjectNotFoundError } from 'infra/http/errors/ObjectNotFoundError';
 import { singleton } from 'tsyringe';
 import { DeleteExamDTO } from '../dtos/DeleteExam.dto';
 import { ExamRepository } from '../repositories/ExamRepository';
@@ -7,6 +8,10 @@ export class DeleteExamService {
   constructor(private examRepository: ExamRepository) {}
 
   async execute({ id }: DeleteExamDTO): Promise<void> {
+    if (!(await this.examRepository.exists(id))) {
+      throw new ObjectNotFoundError('prova', id);
+    }
+
     await this.examRepository.delete(id);
   }
 }
